pyproject.toml
[tool.poetry]
name = "data-platform"
version = "0.1.0"
description = "Data platform for DestaquesGovBr - PostgreSQL migration and data pipelines"
authors = ["Nitai Bezerra <nitai@example.com>"]
readme = "README.md"
packages = [{include = "data_platform", from = "src"}]

[tool.poetry.dependencies]
python = "^3.12"

# Database
psycopg2-binary = "^2.9.9"
sqlalchemy = "^2.0.23"

# HuggingFace
datasets = ">=3.1.0"
huggingface-hub = ">=0.20.0"

# Data processing
pandas = ">=2.1.4"
pyarrow = ">=15.0.0"
numpy = ">=1.26.0"

# HTTP client (Phase 4.7 - Embeddings via Cloud Run API)
httpx = "^0.27.0"
google-auth = "^2.29.0"

# Markdown (for enrichment processing)
markdown = "^3.7"

# Configuration
pydantic = "^2.5.3"
pydantic-settings = "^2.1.0"
python-dotenv = "^1.0.0"
pyyaml = "^6.0.2"

# Retry (used by dataset_manager, enrichment_manager)
retry = "^0.9.2"

# Utilities
tqdm = "^4.66.1"
loguru = "^0.7.2"
typer = "^0.21.0"
typesense = "^1.3.0"
pre-commit = "^4.5.1"

# Workers (Cloud Run services)
fastapi = "^0.115.0"
uvicorn = {version = "^0.34.0", extras = ["standard"]}
textstat = "^0.7.13"
google-cloud-storage = "^3.9.0"
google-cloud-bigquery = "^3.40.1"

[tool.poetry.group.dev.dependencies]
# Testing
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.21.1"
pytest-env = "^1.1.3"
# Code quality
black = "^23.12.1"
ruff = "^0.1.8"
mypy = "^1.7.1"
# Database testing
pytest-postgresql = "^5.0.0"

[tool.poetry.scripts]
data-platform = "data_platform.cli:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py312']
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py312"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "datasets.*",
    "huggingface_hub.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--verbose",
    "--cov=data_platform",
    "--cov-report=term-missing",
    "--cov-report=html",
]
env = [
    "TESTING=1",
]

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
